improved code isolation
