+++ common/schemaTester/async_thread.js
@@ -1,0 +1,68 @@
+
+
+const {
+  Worker, isMainThread, parentPort, threadId
+} = require('worker_threads');
+
+const { async_thread_target_function, threadWork, threadResult } = require("./async_custom");
+
+if (isMainThread) {
+  /**
+   * validates a list of work on mutliple threads
+   * @param {threadWork[]} work 
+   * @param {number} max_threads 
+   * @returns {Promise<threadResult[]>}
+   */
+  module.exports = (work, max_threads) =>
+    new Promise((resolve, reject) => {
+      let threadTotal = Math.min(work.length, max_threads);
+
+      if (!threadTotal) {
+        reject('No work to do.');
+      } else if (threadTotal === 1) {
+        //inline
+        resolve(work.map(w => async_thread_target_function(w, 0)));
+      } else {
+        //multi-thread
+        const expectedResults = work.length;
+
+        /** @type {threadResult[]} */
+        let results = [];
+
+        for (let i = 0; i < threadTotal; i++) {
+          let worker = new Worker(__filename);
+
+          let workHandler = (/** @type {threadResult} */ m) => {
+            if (m) {
+              results.push(m);
+            }
+
+            worker.postMessage(work.pop()); //null work will tell the worker to exit
+
+            if (results.length === expectedResults) {
+              resolve(results);
+            }
+          };
+
+          worker.on('online', workHandler);
+          worker.on('message', workHandler);
+          worker.on('error', reject);
+          worker.on('exit', code => {
+            if (code !== 0)
+              reject(new Error(`Worker stopped with exit code ${code}`));
+          });
+        }
+      }
+    })
+
+} else {
+  parentPort.on("message", (/** @type {*} */ mywork) => {
+    if (mywork) {
+      let result = async_thread_target_function(mywork, threadId);
+
+      parentPort.postMessage(result);
+    } else {
+      process.exit(); //nothing to do...ok to stop thread
+    }
+  });
+}
